wip: genesis and mining for filcns. shardActor support pending
